Version 3.0.0: Tidy up and remove display logic
This commit tidies up the project, removing references to an old version
which logged to SQLite and served web pages.

This project now *only* logs temperatures (to InfuxDB). It does not
display them. For temperature display, have a look at
euoia/temperature-display on GitHub.

The readme has been updated with more information.